[efi] Mark Ip4Config.h as a non-imported header
The Ip4Config.h header has been removed from the EDK2 codebase as
obsolete.  However, we may still encounter it in the wild and so it is
useful to retain the GUID and the corresponding protocol name for
debug messages.

Add an iPXE include guard to this file so that the EDK2 header import
script will no longer attempt to import it from the EDK2 tree.

Signed-off-by: Michael Brown <mcb30@ipxe.org>